+++ CHANGELOG.md
@@ -53,7 +53,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Setting to use the first image as the poster for parent folders [`850ec8b`](https://github.com/ollm/OpenComic/commit/850ec8b83647402d98ea8075e7a020f0bdbcd397)
 - Migrate from `epubjs` to `foliatejs` and update related functionalities [`074de24`](https://github.com/ollm/OpenComic/commit/074de2422c8ff56c4a5eb7bb0e4f25bf1b916afb)
 - Support for `MOBI`, `AZW`, `AZW3` and `FB2` formats [`cdfbebe`](https://github.com/ollm/OpenComic/commit/cdfbebebbf0fa0e6699a2e799d4c4e6218ab1b89)
-
+- Added support for custom themes
 
 ##### 🐛 Bug Fixes